+++ src/kubelet/metric/cadvisor.go
@@ -19,6 +19,16 @@ const (
 	StandaloneCAdvisorMetricsPath = "/metrics"
 )
 
+// getLabel returns the first label it finds by the given names
+func getLabel(labels prometheus.Labels, names ...string) (string, bool) {
+	for _, name := range names {
+		if labels.Has(name) {
+			return labels[name], true
+		}
+	}
+	return "", false
+}
+
 // CadvisorFetchFunc creates a FetchFunc that fetches data from the kubelet cadvisor metrics path.
 func CadvisorFetchFunc(c client.HTTPClient, queries []prometheus.Query) data.FetchFunc {
 	return func() (definition.RawGroups, error) {
@@ -34,54 +44,52 @@ func CadvisorFetchFunc(c client.HTTPClient, queries []prometheus.Query) data.Fet
 		}
 
 		for _, f := range families {
-			if f.Name == "container_memory_usage_bytes" {
-				for _, m := range f.Metrics {
+			for _, m := range f.Metrics {
 
-					// containerName is used for generating the rawEntityID
-					containerName := m.Labels["container_name"]
-					if containerName == "POD" {
-						// skipping metrics from pod containers
-						continue
-					}
+				if label, _ := getLabel(m.Labels, "container_name", "container"); label == "POD" {
+					// skipping metrics from pod containers
+					continue
+				}
 
-					if containerName == "" {
-						errs = append(errs, errors.New("container name not found in cadvisor metrics"))
-						continue
-					}
+				rawEntityID, err := createRawEntityID(m)
+				if err != nil {
+					errs = append(errs, err)
+					continue
+				}
 
-					// namespace is used for generating the rawEntityID
-					namespace := m.Labels["namespace"]
-					if namespace == "" {
-						errs = append(errs, errors.New("namespace not found in cadvisor metrics"))
-						continue
-					}
+				// It does not belong to a container that we care about. Special case that does not warrant an error
+				if rawEntityID == "" {
+					continue
+				}
 
-					// namespace is used for generating the rawEntityID
-					podName := m.Labels["pod_name"]
-					if podName == "" {
-						errs = append(errs, errors.New("pod name not found in cadvisor metrics"))
-						continue
-					}
+				containerID := extractContainerID(m.Labels["id"])
+				if containerID == "" {
+					errs = append(errs, errors.New("container id not found in cAdvisor metrics"))
+					continue
+				}
 
-					rawEntityID := fmt.Sprintf("%s_%s_%s", namespace, podName, containerName)
+				var (
+					metrics definition.RawMetrics
+					ok      bool
+				)
 
-					container := make(definition.RawMetrics, 2)
+				if metrics, ok = g["container"][rawEntityID]; !ok {
+					metrics = make(definition.RawMetrics)
+					metrics["containerID"] = containerID
+					g["container"][rawEntityID] = metrics
+				}
 
-					if v := extractContainerID(m.Labels["id"]); v == "" {
-						errs = append(errs, errors.New("container id not found in cadvisor metrics"))
-						continue
-					} else {
-						container["containerID"] = v
-					}
-
+				switch f.Name {
+				case "container_memory_usage_bytes":
+					// Special case, where we are only interested in collecting containerId and containerImageID
 					if m.Labels["image"] == "" {
-						errs = append(errs, errors.New("container image not found in cadvisor metrics"))
+						errs = append(errs, errors.New("container image not found in cAdvisor metrics"))
 						continue
 					}
-
-					container["containerImageID"] = m.Labels["image"]
-
-					g["container"][rawEntityID] = container
+					metrics["containerImageID"] = m.Labels["image"]
+				default:
+					// by default, we want the actual metric
+					metrics[f.Name] = m.Value
 				}
 			}
 		}
@@ -95,6 +103,29 @@ func CadvisorFetchFunc(c client.HTTPClient, queries []prometheus.Query) data.Fet
 
 		return g, nil
 	}
+}
+
+func createRawEntityID(m prometheus.Metric) (string, error) {
+	containerName, ok := getLabel(m.Labels, "container_name", "container")
+	if !ok {
+		return "", errors.New("container name not found in cAdvisor metrics")
+	}
+	if containerName == "" {
+		return "", nil
+	}
+
+	namespace := m.Labels["namespace"]
+	if namespace == "" {
+		return "", errors.New("namespace not found in cAdvisor metrics")
+
+	}
+
+	podName, _ := getLabel(m.Labels, "pod_name", "pod")
+	if podName == "" {
+		return "", errors.New("pod name not found in cAdvisor metrics")
+	}
+
+	return fmt.Sprintf("%s_%s_%s", namespace, podName, containerName), nil
 }
 
 // /kubepods/besteffort/podba8b34d7-11a3-11e8-a084-080027352a02/a949bd136c1397b9f52905538ee11450427be33648abe38db06be2e5cfbeca49