+++ src/kubernetes.go
@@ -68,7 +68,7 @@ const (
 	defaultDiscoveryCacheTTL           = time.Hour
 
 	integrationName    = "com.newrelic.kubernetes"
-	integrationVersion = "1.22.0"
+	integrationVersion = "1.23.0"
 	nodeNameEnvVar     = "NRK8S_NODE_NAME"
 )